+++ test/util.h
@@ -12,7 +12,7 @@ static std::string GenRandomString(size_t len) {
     std::string result;
     result.reserve(len);
     for (size_t i = 0; i < len; ++i) {
-        result += alphanum[std::rand() % (sizeof(alphanum) - 1)];
+        result += alphanum[static_cast<size_t>(std::rand()) % (sizeof(alphanum) - 1)];
     }
     return result;
 }
